refactor: 체스 컨트롤러 제거 및 스프링 컨트롤러에 통합
